refactor: move fixImport function to new registry-utils
fix: replace composables instead of hooks in regex
